Refactor DatabaseManager: Remove metrics summary methods and integrate AnalyticsRepo
- Removed upsert_metrics_summary, get_metrics_summary, get_heavy_summary, and get_deleted_tallies methods.
- Added AnalyticsRepo instance to DatabaseManager for analytics-related operations.
- Cleaned up unused imports and code related to metrics summary.
